Fix up toggle button
